requirements.txt
flask
requests
Pillow
gunicorn
python-magic